pyproject.toml
[project]
name = "adidas-image-match"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"faiss-cpu>=1.8.0",
	"numpy>=2.0.0",
	"openai>=1.40.0",
	"pillow>=10.0.0",
	"python-dotenv>=1.0.0",
	"pyyaml>=6.0.0",
	"requests>=2.32.0",
	"timm>=1.0.0",
	"torch>=2.3.0",
]

[project.scripts]
run_reference_indexing.py = "src.pipeline.run_reference_indexing:main"
